Changed object IDs to type Long.

+++ src/main/java/org/yesworkflow/extract/Source.java
@@ -4,17 +4,16 @@ import org.yesworkflow.db.YesWorkflowDB;
 
 public class Source {
 
-
-    public final Integer id;
+    public final Long id;
     public final String path;
     
     public static Source newSource(YesWorkflowDB ywdb, String path) {
-        int id = ywdb.insertSourceFile(path);
+        Long id = ywdb.insertSourceFile(path);
         Source source = new Source(id, path);
         return source;
     }
     
-    public Source(Integer id, String path) {
+    public Source(Long id, String path) {
         this.id = id;
         this.path = path;
     }
